Refactoring to reduce compiled program size. Lots of things broken at the moment.

+++ PlayerShip.h
@@ -7,7 +7,7 @@ public:
 	PlayerShip();
 	void update();
 	void render(Vector2Int cameraPos);
-	float getCameraTarget();
+	int getCameraTarget();
 	void destroy();
 	Rect getCollisionRect();
 	byte explosionTimer = 0;
